+++ src/Query/CallbackBuilder.php
@@ -14,6 +14,7 @@ use PhpParser\Node\Stmt\Expression;
 use PhpParser\Node\Stmt\Return_;
 use PhpParser\PrettyPrinter\Standard;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
+use Remorhaz\JSON\Path\Runtime\EvaluatorInterface;
 use Remorhaz\JSON\Path\Value\ValueListInterface;
 use Remorhaz\JSON\Path\Runtime\RuntimeInterface;
 use Remorhaz\UniLex\AST\AbstractTranslatorListener;
@@ -26,6 +27,8 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
 
     private const ARG_RUNTIME = 'runtime';
 
+    private const ARG_EVALUATOR = 'evaluator';
+
     private const ARG_INPUT = 'input';
 
     private $php;
@@ -33,6 +36,8 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
     private $references = [];
 
     private $runtime;
+
+    private $evaluator;
 
     private $input;
 
@@ -68,20 +73,26 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
     public function onStart(QueryAstNode $node): void
     {
         $this->runtime = $this->php->var(self::ARG_RUNTIME);
+        $this->evaluator = $this->php->var(self::ARG_EVALUATOR);
         $this->input = $this->php->var(self::ARG_INPUT);
     }
 
     public function onFinish(): void
     {
+        $inputParam = $this
+            ->php
+            ->param(self::ARG_INPUT)
+            ->setType(NodeValueInterface::class)
+            ->getNode();
         $runtimeParam = $this
             ->php
             ->param(self::ARG_RUNTIME)
             ->setType(RuntimeInterface::class)
             ->getNode();
-        $inputParam = $this
+        $evaluatorParam = $this
             ->php
-            ->param(self::ARG_INPUT)
-            ->setType(NodeValueInterface::class)
+            ->param(self::ARG_EVALUATOR)
+            ->setType(EvaluatorInterface::class)
             ->getNode();
         $stmts = array_map(
             function (PhpAstNode $stmt): PhpAstNode {
@@ -94,7 +105,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
             [
                 'stmts' => $stmts,
                 'returnType' => ValueListInterface::class,
-                'params' => [$runtimeParam, $inputParam],
+                'params' => [$inputParam, $runtimeParam, $evaluatorParam],
             ]
         );
         $return = new Return_($closure);
@@ -119,7 +130,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
         }
         switch ($node->getName()) {
             case AstNodeType::GET_INPUT:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'getInput',
                     $this->input
@@ -135,7 +146,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::CREATE_FILTER_CONTEXT:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'createFilterContext',
                     $this->getReference($node->getChild(0))
@@ -143,7 +154,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::SPLIT:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'split',
                     $this->getReference($node->getChild(0))
@@ -151,7 +162,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::EVALUATE:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
                     'evaluate',
                     $this->getReference($node->getChild(0)),
@@ -160,7 +171,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::FILTER:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'filter',
                     $this->getReference($node->getChild(0)),
@@ -169,60 +180,60 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::EVALUATE_LOGICAL_OR:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'evaluateLogicalOr',
+                    'logicalOr',
                     $this->getReference($node->getChild(0)),
                     $this->getReference($node->getChild(1))
                 );
                 break;
 
             case AstNodeType::EVALUATE_LOGICAL_AND:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'evaluateLogicalAnd',
+                    'logicalAnd',
                     $this->getReference($node->getChild(0)),
                     $this->getReference($node->getChild(1))
                 );
                 break;
 
             case AstNodeType::EVALUATE_LOGICAL_NOT:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'evaluateLogicalNot',
+                    'logicalNot',
                     $this->getReference($node->getChild(0))
                 );
                 break;
 
             case AstNodeType::CALCULATE_IS_EQUAL:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'calculateIsEqual',
+                    'isEqual',
                     $this->getReference($node->getChild(0)),
                     $this->getReference($node->getChild(1))
                 );
                 break;
 
             case AstNodeType::CALCULATE_IS_GREATER:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'calculateIsGreater',
+                    'isGreater',
                     $this->getReference($node->getChild(0)),
                     $this->getReference($node->getChild(1))
                 );
                 break;
 
             case AstNodeType::CALCULATE_IS_REGEXP:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
-                    'calculateIsRegExp',
+                    'isRegExp',
                     $this->php->val($node->getAttribute('pattern')),
                     $this->getReference($node->getChild(0))
                 );
                 break;
 
             case AstNodeType::FETCH_CHILDREN:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'fetchChildren',
                     $this->getReference($node->getChild(0)),
@@ -235,7 +246,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::FETCH_CHILDREN_DEEP:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'fetchChildrenDeep',
                     $this->getReference($node->getChild(0)),
@@ -248,7 +259,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::MATCH_ANY_CHILD:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'matchAnyChild',
                     $this->getReference($node->getChild(0))
@@ -256,7 +267,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::MATCH_PROPERTY_STRICTLY:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'matchPropertyStrictly',
                     $this->getReference($node->getChild(0)),
@@ -264,7 +275,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::MATCH_ELEMENT_STRICTLY:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'matchElementStrictly',
                     $this->getReference($node->getChild(0)),
@@ -272,7 +283,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::AGGREGATE:
-                $this->addMethodCall(
+                $this->addEvaluatorMethodCall(
                     $node,
                     'aggregate',
                     $this->php->val($node->getAttribute('name')),
@@ -281,7 +292,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::POPULATE_LITERAL:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'populateLiteral',
                     $this->getReference($node->getChild(0)),
@@ -290,7 +301,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::POPULATE_ARRAY_ELEMENTS:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'populateArrayElements',
                     $this->getReference($node->getChild(0)),
@@ -303,7 +314,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::POPULATE_INDEX_LIST:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'populateIndexList',
                     $this->getReference($node->getChild(0)),
@@ -316,7 +327,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
 
             case AstNodeType::POPULATE_INDEX_SLICE:
                 $attributes = $node->getAttributeList();
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'populateIndexSlice',
                     $this->getReference($node->getChild(0)),
@@ -328,7 +339,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::POPULATE_NAME_LIST:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'populateNameList',
                     $this->getReference($node->getChild(0)),
@@ -343,7 +354,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 $attributes = $node->getAttributeList();
                 // TODO: allow accessing null attributes
                 $value = $this->php->val($attributes['value'] ?? null);
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'createScalar',
                     $value
@@ -371,7 +382,7 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
                 break;
 
             case AstNodeType::CREATE_LITERAL_ARRAY:
-                $this->addMethodCall(
+                $this->addRuntimeMethodCall(
                     $node,
                     'createArray',
                     $this->getReference($node->getChild(0)),
@@ -421,11 +432,19 @@ final class CallbackBuilder extends AbstractTranslatorListener implements Callba
         return $this->references[$node->getId()];
     }
 
-    private function addMethodCall(QueryAstNode $node, string $method, PhpAstNode ...$args): void
+    private function addRuntimeMethodCall(QueryAstNode $node, string $method, PhpAstNode ...$args): void
     {
         $methodCall = $this
             ->php
             ->methodCall($this->runtime, $method, $args);
+        $this->stmts[] = new Assign($this->createReference($node), $methodCall);
+    }
+
+    private function addEvaluatorMethodCall(QueryAstNode $node, string $method, PhpAstNode ...$args): void
+    {
+        $methodCall = $this
+            ->php
+            ->methodCall($this->evaluator, $method, $args);
         $this->stmts[] = new Assign($this->createReference($node), $methodCall);
     }
 }