+++ src/Query/Query.php
@@ -5,6 +5,7 @@ namespace Remorhaz\JSON\Path\Query;
 
 use function call_user_func;
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
+use Remorhaz\JSON\Path\Runtime\EvaluatorInterface;
 use Remorhaz\JSON\Path\Value\ValueListInterface;
 use Remorhaz\JSON\Path\Runtime\RuntimeInterface;
 
@@ -24,9 +25,12 @@ final class Query implements QueryInterface
         $this->properties = $properties;
     }
 
-    public function __invoke(RuntimeInterface $runtime, NodeValueInterface $rootNode): ValueListInterface
-    {
-        return call_user_func($this->callback, $runtime, $rootNode);
+    public function __invoke(
+        NodeValueInterface $rootNode,
+        RuntimeInterface $runtime,
+        EvaluatorInterface $evaluator
+    ): ValueListInterface {
+        return call_user_func($this->callback, $rootNode, $runtime, $evaluator);
     }
 
     public function getCapabilities(): CapabilitiesInterface