+++ src/Query/QueryInterface.php
@@ -4,13 +4,18 @@ declare(strict_types=1);
 namespace Remorhaz\JSON\Path\Query;
 
 use Remorhaz\JSON\Data\Value\NodeValueInterface;
+use Remorhaz\JSON\Path\Runtime\EvaluatorInterface;
 use Remorhaz\JSON\Path\Value\ValueListInterface;
 use Remorhaz\JSON\Path\Runtime\RuntimeInterface;
 
 interface QueryInterface
 {
 
-    public function __invoke(RuntimeInterface $runtime, NodeValueInterface $rootNode): ValueListInterface;
+    public function __invoke(
+        NodeValueInterface $rootNode,
+        RuntimeInterface $runtime,
+        EvaluatorInterface $evaluator
+    ): ValueListInterface;
 
     public function getCapabilities(): CapabilitiesInterface;