+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Changed
+- Evaluator is no more part of runtime object.
+- Query invocation signature changed.
+
 ## [0.2.0] - 2019-07-28
 ### Added
 - Processor supports selecting value paths (single and multiple).